requirements.txt
pygame-ce>=2.4.0
pillow>=10.0.0
